Update environment/client configuration loading
This update extends the boto3 client adapter to accept connections
beyond Kinesis streams, and moves it to a helper function, where
it can be easily imported to be used anywhere in a function.

The environment file loader has also been made a standalone helper
function.

Beyond these updates, this commit adds a new 'local' development
environment that allows different variables to be set for a local
environment from the development environment. (Previously they were
assumed to be the same)
