+++ requirements.txt
@@ -1,3 +1,4 @@
+boto3
 python-lambda
 coverage
 flake8
